requirements.txt
annotated-types==0.7.0
anyio==4.5.0
blinker==1.9.0
blis==1.3.0
CacheControl==0.14.2
cachetools==5.5.2
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
cloudpathlib==0.21.0
colorama==0.4.6
confection==0.1.5
cryptography==44.0.2
cymem==2.0.11
dnspython==2.6.1
email_validator==2.2.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
fastapi==0.115.0
fastapi-cli==0.0.5
firebase-admin==6.7.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.1
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
langcodes==3.5.0
language_data==1.3.0
lxml==5.3.2
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
msgpack==1.1.0
murmurhash==1.0.12
nltk==3.9.1
numpy==2.2.2
packaging==24.2
pandas==2.2.3
preshed==3.0.9
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.10.1
pyparsing==3.2.3
PyPDF2==3.0.1
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytube==15.0.0
pytz==2025.1
PyYAML==6.0.2
qrcode==7.4.2
regex==2024.11.6
requests==2.32.3
rich==13.8.1
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.1
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
spacy==3.8.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.40
srsly==2.5.1
starlette==0.38.5
thinc==8.3.6
threadpoolctl==3.5.0
tqdm==4.67.1
typer==0.12.5
typing_extensions==4.10.0
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.30.6
wasabi==1.1.3
watchfiles==0.24.0
weasel==0.4.1
websockets==13.0.1
Werkzeug==3.1.3
wrapt==1.17.2
